+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## UNRELEASED (YYYY-MM-DD)
 
+### Added
+
+- the `sea_orm` crate feature allows to use `MagnetLink`, `TorrentFile` and `TorrentID`
+  in sea_orm models
+- `TorrentFile`, `DecodedTorrent` and `DecodedInfo` implement `PartialEq`
+- `MagnetFile` implements `FromStr`, `TryFrom<String>`, and `Into<String>`
+- `MagnetFile` supports serde de/serialization
+
 ## Version 0.4.0 (2025-11-10)
 
 This release is focused on stricter parsing of torrents and magnets, and